% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCodonFrequency.R
\name{check_fasta_codon_bias}
\alias{check_fasta_codon_bias}
\title{Codon Frequency}
\usage{
check_fasta_codon_bias(fasta_path, write_bedgraph = TRUE)
}
\arguments{
\item{fasta_path}{path to fasta file (string)}

\item{write_bedgraph}{write bedgraph files for each fasta sequence and reading frame? (bool)}
}
\value{
table describing frequency of each codon in the fasta, as calculated using the entire coding genome (dataframe)
}
\description{
Take a fasta file and return a dataframe describing how common its codons are in the human genome.
}
