% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCodonFrequency.R
\name{codon_freq}
\alias{codon_freq}
\title{Codon Frequency}
\usage{
codon_freq()
}
\value{
dataframe describing pre-calculated rates at which codons occur in the coding genome
}
\description{
Codon Frequency
}
\examples{
codon_freq()
}
